Consolidate price helpers: rename to price-helpers.ts, unify ZERO_PRICE, move costAsPrice
- Renamed zero-price.ts → price-helpers.ts (better describes purpose)
- Unified ZERO_PRICE: removed separate TEXT/IMAGE/ALL variants
- Moved costAsPrice helper from text.ts and image.ts to price-helpers.ts
- Made costAsPrice generic to work with any model provider registry
- Updated all imports and usages in text.ts and image.ts

Benefits:
- Single source of truth for all price utilities
- One ZERO_PRICE constant works for all model types
- Reusable costAsPrice helper eliminates duplication
- Cleaner, more maintainable code structure
